+++ cpp/libtts-client/Client.h
@@ -1,16 +1,16 @@
-#ifndef __TRIBUNE_CLIENT_H__
-#define __TRIBUNE_CLIENT_H__
+#ifndef TECHMO__TTS__CLIENT__H
+#define TECHMO__TTS__CLIENT__H
 
 #include <optional>
 #include <string>
 #include <string_view>
 #include <vector>
-#include "tribune_tts.grpc.pb.h"
+#include "techmo_tts.grpc.pb.h"
 
-namespace techmo::tribune
+namespace techmo::tts
 {
 	// Desfines the parameters of synthesized speech.
-	struct SynthesizeAudioConfig
+	struct ClientAudioConfig
 	{
 		// Encoding of the output data stream. Either PCM16 without any header or Ogg/Vorbis.
 		AudioEncoding encoding{ AudioEncoding::PCM16 };
@@ -33,7 +33,7 @@ namespace techmo::tribune
 
 	// Voice definition used to describe requested voice in SynthesizeConfig,
 	// and voice parameters in ListVoices (as a part of VoiceInfo).
-	struct SynthesizeVoice
+	struct ClientVoice
 	{
 		// The name of the voice.
 		std::string name;
@@ -47,28 +47,28 @@ namespace techmo::tribune
 
 	// Information about a voice, returned by ListVoices call.
 	// Both the supported_languages and voice are always defined, i.e. non-empy.
-	struct SynthesizeVoiceInfo
+	struct ClientVoiceInfo
 	{
 		// The list of ISO 639-1 codes of laguages supported by the voice.
 		std::vector<std::string> supported_languages;
 
 		// The voice parameters.
-		SynthesizeVoice voice;
+		ClientVoice voice;
 	};
 
-	struct TribuneSynthesizeConfig
+	struct ClientSynthesizeConfig
 	{
 		// ISO 639-1 code of the language of text to be synthesized (optional, may be overriden by SSML tags in request text).
 		std::string language;
 
 		// Parameters to configure the audio synthesis (optional).
-		std::optional<SynthesizeAudioConfig> audio_config;
+		std::optional<ClientAudioConfig> audio_config;
 
 		// Requested voice to be used to synthesize the text (may be overriden by SSML tags in request text).
-		std::optional<SynthesizeVoice> voice;
+		std::optional<ClientVoice> voice;
 	};
 
-	struct TribuneClientConfig
+	struct ClientConfig
 	{
 		// Session ID to be passed to the service. If not specified, the service will generate a default session ID itself.
 		// Session ID is the best way to match log's from client application with these on server side.
@@ -79,7 +79,7 @@ namespace techmo::tribune
 
 	};
 
-	struct TribuneAudioData
+	struct ClientAudioData
 	{
 		// Sample rate in Hz of received audio data.
 		int sample_rate_hertz{ 0 };
@@ -88,26 +88,26 @@ namespace techmo::tribune
 		std::string audio_bytes;
 	};
 
-	class TribuneClient
+	class Client
 	{
 	public:
-		TribuneClient(std::string_view serviceAddress):
+		Client(std::string_view serviceAddress):
 			m_serviceAddress{ serviceAddress }
 		{
 		}
 
-		std::vector<SynthesizeVoiceInfo> ListVoices(
-			const TribuneClientConfig& clientConfig,
+		std::vector<ClientVoiceInfo> ListVoices(
+			const ClientConfig& clientConfig,
 			std::string_view language = "") const;
 
-		TribuneAudioData SynthesizeStreaming(
-			const TribuneClientConfig& clientConfig,
-			const TribuneSynthesizeConfig& synthesizeConfig,
+		ClientAudioData SynthesizeStreaming(
+			const ClientConfig& clientConfig,
+			const ClientSynthesizeConfig& synthesizeConfig,
 			std::string_view text) const;
 
-		TribuneAudioData Synthesize(
-			const TribuneClientConfig& clientConfig,
-			const TribuneSynthesizeConfig& synthesizeConfig,
+		ClientAudioData Synthesize(
+			const ClientConfig& clientConfig,
+			const ClientSynthesizeConfig& synthesizeConfig,
 			std::string_view text) const;
 
 	private:
@@ -116,4 +116,4 @@ namespace techmo::tribune
 
 }
 
-#endif //__TRIBUNE_CLIENT_H__
+#endif