+++ python/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Tribune TTS gRPC Python client Changelog
+# Techmo TTS gRPC Python client Changelog
 
 ## [2.0.0] - 2020-11-04
 ### Added